+++ samplesheets/rules.py
@@ -30,7 +30,7 @@ rules.add_perm(
 rules.add_perm(
     'samplesheets.edit_sheet',
     rules.is_superuser | pr_rules.is_project_owner |
-    pr_rules.is_project_delegate)
+    pr_rules.is_project_delegate | pr_rules.is_project_contributor)
 
 # Allow exporting a JSON sample sheet from project
 rules.add_perm(